+++ src/RazorHx.Htmx/HtmxTrigger.cs
@@ -1,0 +1,8 @@
+namespace RazorHx.Htmx;
+
+public class HtmxTrigger
+{
+    public required string Name { get; init; }
+    public TriggerTiming Timing { get; init; }
+    public object? Parameters { get; init; }
+}