+++ src/RazorHx.Htmx/TriggerTiming.cs
@@ -1,0 +1,8 @@
+namespace RazorHx.Htmx;
+
+public enum TriggerTiming
+{
+    Default,
+    AfterSettle,
+    AfterSwap
+}