+++ src/RazorHx/Results/RazorHxResultExecutor.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using RazorHx.Components;
 using RazorHx.DependencyInjection;
+using RazorHx.Htmx;
 using RazorHx.Htmx.HttpContextFeatures;
 
 namespace RazorHx.Results;
@@ -13,30 +14,16 @@ internal static class RazorHxResultExecutor
 {
     private const string DefaultContentType = "text/html; charset=utf-8";
 
-    public static Task ExecuteAsync(HttpContext httpContext, RazorHxResult result)
+    public static async Task ExecuteAsync(HttpContext httpContext, RazorHxResult result)
     {
         ArgumentNullException.ThrowIfNull(httpContext);
 
         var response = httpContext.Response;
         response.ContentType = result.ContentType ?? DefaultContentType;
+        response.StatusCode = result.StatusCode ?? response.StatusCode;
 
-        if (result.StatusCode != null) response.StatusCode = result.StatusCode.Value;
+        PrepareResponseHeaders(httpContext, result);
 
-        return RenderComponentToResponse(
-            httpContext,
-            result.ComponentType,
-            result.Parameters,
-            result.OobComponentType,
-            result.OobParameters);
-    }
-
-    private static async Task<Task> RenderComponentToResponse(
-        HttpContext httpContext,
-        Type componentType,
-        ParameterView componentParameters,
-        Type? oobComponentType,
-        ParameterView oobComponentParameters)
-    {
         var htmlRenderer = httpContext.RequestServices.GetRequiredService<HtmlRenderer>();
         var razorHxComponentsServiceOptions = httpContext.RequestServices.GetRequiredService<RazorHxServiceOptions>();
         var htmxRequestFeature = httpContext.Features.Get<IHtmxRequestFeature>();
@@ -52,23 +39,23 @@ internal static class RazorHxResultExecutor
         }
         else
         {
-            var useComponentLayout = componentType.GetCustomAttribute<LayoutAttribute>() != null;
+            var useComponentLayout = result.ComponentType.GetCustomAttribute<LayoutAttribute>() != null;
             layout = useComponentLayout
-                ? componentType.GetCustomAttribute<LayoutAttribute>()!.LayoutType
+                ? result.ComponentType.GetCustomAttribute<LayoutAttribute>()!.LayoutType
                 : razorHxComponentsServiceOptions.RootComponent;
         }
 
         var parameters = new Dictionary<string, object?>
         {
             { "Layout", layout },
-            { "ComponentType", componentType },
-            { "Parameters", (Dictionary<string, object?>)componentParameters.ToDictionary() }
+            { "ComponentType", result.ComponentType },
+            { "Parameters", (Dictionary<string, object?>)result.Parameters.ToDictionary() }
         };
 
-        if (oobComponentType != null)
+        if (result.OobComponentType != null)
         {
-            parameters.Add("OobComponentType", oobComponentType);
-            parameters.Add("OobParameters", (Dictionary<string, object?>)oobComponentParameters.ToDictionary());
+            parameters.Add("OobComponentType", result.OobComponentType);
+            parameters.Add("OobParameters", (Dictionary<string, object?>)result.OobParameters.ToDictionary());
         }
 
         var htmlContent = await htmlRenderer.Dispatcher.InvokeAsync(async () =>
@@ -78,6 +65,41 @@ internal static class RazorHxResultExecutor
             return output.ToHtmlString();
         });
 
-        return httpContext.Response.WriteAsync(htmlContent);
+        await httpContext.Response.WriteAsync(htmlContent);
+    }
+
+    private static void PrepareResponseHeaders(HttpContext httpContext, RazorHxResult result)
+    {
+        foreach (var trigger in result.Triggers)
+        {
+            switch (trigger.Timing)
+            {
+                case TriggerTiming.Default:
+                    SetHeader(HxResponseHeaderKeys.Trigger, trigger.Name);
+                    break;
+                case TriggerTiming.AfterSettle:
+                    SetHeader(HxResponseHeaderKeys.TriggerAfterSettle, trigger.Name);
+                    break;
+                case TriggerTiming.AfterSwap:
+                    SetHeader(HxResponseHeaderKeys.TriggerAfterSwap, trigger.Name);
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+        }
+
+        return;
+
+        void SetHeader(string header, string trigger)
+        {
+            if (httpContext.Response.Headers.Remove(header, out var _))
+            {
+                httpContext.Response.Headers.Append(header, trigger);
+            }
+            else
+            {
+                httpContext.Response.Headers.Append(header, trigger);
+            }
+        }
     }
 }