+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.3.0] - 2025-07-11
+
+### Added
+
+- HttpContext extensions to check htmx request headers
+- Add a Trigger to the RazorHxResult.
+
 ## [0.2.0] - 2025-01-13
 
 ### Added